fix codesnip module structure
