+++ crates/codesnip_core/Cargo.toml
@@ -4,9 +4,6 @@ version = "0.1.0"
 authors = ["to-omer <toomerhs@gmail.com>"]
 edition = "2018"
 
-[lib]
-doc = false
-
 [dependencies]
 syn = { version = "1", features = ["full", "extra-traits", "visit", "visit-mut"] }
 proc-macro2 = "1"
