+++ crates/codesnip_core/src/lib.rs
@@ -1,5 +1,10 @@
 pub mod entry;
-pub mod ext;
-pub mod format;
-pub mod map;
-pub mod parse;
+mod ext;
+mod format;
+mod map;
+mod parse;
+
+pub use ext::{AttributeExt, ItemExt, PathExt};
+pub use format::{format_with_rustfmt, rustfmt_exits};
+pub use map::{Filter, LinkedSnippet, SnippetMap};
+pub use parse::{parse_file_recursive, Error};
